Add Support for Returning Status Code and Reason
It's common for agents to incorrectly use a tool at first, which can be
handled if is able to see the status code and reason information from
the server. Calling `.text` on everything strips that information.

This PR doesn't alter the defaults but rather lets the user add a format
template similar to common log syntax.